* adds mechanism for revoking JWT tokens

+++ src/user-session/session-token.service.ts
@@ -1,15 +1,19 @@
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
-  UnauthorizedException
+  UnauthorizedException,
 } from '@nestjs/common';
-
+import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
+import { InjectModel } from '@nestjs/sequelize';
 
-import { TokenType, SessionToken } from './session-token.entity';
+import { SessionToken } from './session-token.entity';
+
+import { RedisService } from '../redis/redis.service';
 
 import * as dayjs from 'dayjs';
-import { InjectModel } from '@nestjs/sequelize';
+
 
 export type JwtTokens = { accessToken: string; refreshToken: string };
 
@@ -17,8 +21,15 @@ export type JwtTokens = { accessToken: string; refreshToken: string };
 export class SessionTokenService {
   constructor(
     @InjectModel(SessionToken) private readonly sessions: typeof SessionToken,
-    private readonly jwtService: JwtService
+    private readonly configService: ConfigService,
+    private readonly jwtService: JwtService,
+    private readonly redis: RedisService
   ) {}
+
+  private isTokenExpired(expiresAt: Date): boolean {
+    return dayjs().isAfter(dayjs(expiresAt));
+  }
+
   async addRefreshToken(
     userId: number,
     refreshToken: string
@@ -26,7 +37,6 @@ export class SessionTokenService {
     return await this.sessions.create({
       expiresAt: dayjs().add(7, 'day').toDate(),
       refreshToken,
-      typeOfToken: TokenType.REFRESH,
       userId
     });
   }
@@ -35,6 +45,10 @@ export class SessionTokenService {
     userId: number,
     refreshToken: string
   ): Promise<SessionToken> {
+    if (await this.redis.getKey(refreshToken)) {
+      throw new ForbiddenException('Token has been revoked.');
+    }
+
     const session = await this.sessions.findOne({
       raw: true,
       where: { userId, refreshToken }
@@ -44,7 +58,7 @@ export class SessionTokenService {
       throw new UnauthorizedException('Invalid refresh token.');
     }
 
-    if (dayjs().isAfter(dayjs(session?.expiresAt))) {
+    if (this.isTokenExpired(session?.expiresAt)) {
       throw new ForbiddenException('Refresh token has expired.');
     }
 
@@ -55,20 +69,41 @@ export class SessionTokenService {
     userId: number,
     refreshToken: string
   ): Promise<void> {
-    await this.sessions.update(
-      {
-        deletedAt: dayjs().toDate()
-      },
-      {
-        where: { userId, refreshToken }
-      }
+    const session = await this.sessions.findOne({
+      where: { userId, refreshToken }
+    });
+    if (!session) {
+      throw new BadRequestException('Refresh token not found.');
+    }
+
+    await this.redis.setKey(
+      refreshToken,
+      userId.toString(),
+      dayjs(session.expiresAt).diff(dayjs(), 'seconds')
     );
+
+    return await session.destroy();
   }
 
-  async removeAllRefreshTokens(userId: number): Promise<void> {
-    await this.sessions.destroy({
+  async removeAllRefreshTokens(userId: number): Promise<number> {
+    const sessions = await this.sessions.findAll({
+      raw: true,
       where: { userId }
     });
+
+    if (!sessions.length) {
+      return 0;
+    }
+
+    for (const session of sessions) {
+      await this.redis.setKey(
+        session.refreshToken,
+        userId.toString(),
+        dayjs(session.expiresAt).diff(dayjs(), 'second')
+      );
+    }
+
+    return await this.sessions.destroy({ where: { userId } });
   }
 
   async refreshToken(
@@ -80,10 +115,13 @@ export class SessionTokenService {
     }
 
     // Generate new access & refresh tokens
-    const newAccessToken = this.jwtService.sign({ userId });
+    const newAccessToken = this.jwtService.sign(
+      { sub: userId },
+      { expiresIn: this.configService.get('JWT_ACCESS_TOKEN_EXPIRES') }
+    );
     const newRefreshToken = this.jwtService.sign(
-      { userId },
-      { expiresIn: '7d' }
+      { sub: userId },
+      { expiresIn: this.configService.get('JWT_REFRESH_TOKEN_EXPIRES') }
     );
 
     // Store new refresh token and remove the old one
@@ -93,8 +131,14 @@ export class SessionTokenService {
   }
 
   async generateTokens(userId: number): Promise<JwtTokens> {
-    const accessToken = this.jwtService.sign({ userId }, { expiresIn: '15m' });
-    const refreshToken = this.jwtService.sign({ userId }, { expiresIn: '7d' });
+    const accessToken = this.jwtService.sign(
+      { sub: userId },
+      { expiresIn: this.configService.get('JWT_ACCESS_TOKEN_EXPIRES') }
+    );
+    const refreshToken = this.jwtService.sign(
+      { sub: userId },
+      { expiresIn: this.configService.get('JWT_REFRESH_TOKEN_EXPIRES') }
+    );
 
     await this.addRefreshToken(userId, refreshToken);
